% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join_flip2sss.R
\name{join_flip2sss}
\alias{join_flip2sss}
\title{join tests from multiverse flip2sss models}
\usage{
join_flip2sss(formulas, summstats_within, data, cluster, tested_coeffs = NULL, n_flips = 5000, flips = NULL, seed=NULL, ...)
}
\arguments{
\item{formulas}{A formula or a list of formulas. It can be a (list) of complete models as.formula or a list of formulas}

\item{summstats_within}{A (list of) vector of summary statistics model within the data or a (list of) function with argument data.}

\item{data}{The dataset to be used for fitting the model.}

\item{cluster}{A vector or a formula evaluated on the data that defines the clusters.}

\item{tested_coeffs}{list of the same length of \code{mods}, each element of the list being a vector of
names of tested coefficients. Alternatively, it can be a vector of names of tested coefficients, in this case, the tested coefficients are attributed to all models (when present). 
As a last option, it can be \code{NULL}, if so, all coefficients are tested.}

\item{n_flips}{number of flips, default 5000}

\item{flips}{matrix fo +1 or -1, the matrix has \code{n_flips} rows and n (number of observations) columns}

\item{seed}{\code{NULL} by default. Use a number if you wanto to ensure replicability of the results}

\item{...}{any other further parameter.}
}
\value{
A \code{jointest} object, i.e., a list containing the following objects: 
\describe{
 \item{Tspace}{\code{data.frame} where rows represents the sign-flipping transformed (plus the identity one) test and columns the variables.}
 \item{summary_table}{\code{data.frame} containing for each model the estimated parameter(s), score(s), std error(s), test(s), partial correlation(s) and p-value(s).} 
 \item{mods}{List of \code{glm}s or \code{flipscores} objects.}
}
}
\description{
The function allows hypothesis testing across all plausible multiverse flip2sss models
ensuring strong family-wise error rate control.
}
\examples{
N=20
n=rpois(N,20)
reff=rep(rnorm(N),n)
D=data.frame(X1=rnorm(length(reff)),
               X2=rep(rnorm(N),n),
               Grp=factor(rep(rep(LETTERS[1:3],length.out=N),n)),
               Subj=rep(1:N,n))
D$Y=rbinom(n=nrow(D),prob=plogis( 2*D$X1 * (D$Grp=="B") +  2*D$X2+reff),size=1)
summstats_within <- list('glm(Y ~ X1, family = binomial(link = "logit"))',
                        'logistf::logistf(Y ~ X1^2, family = binomial(link = "logit"), control=logistf::logistf.control(maxit=100))')
formulas <- list(Y ~ Grp * X1 + X2, Y ~ Grp + X1 + X2,  Y ~ Grp + X1)
res <- join_flip2sss(formulas = formulas, summstats_within = summstats_within, 
                   data = D, cluster = D$Subj) 
}
