% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R, R/padjust.R
\docType{methods}
\name{jointest-method}
\alias{jointest-method}
\alias{summary.jointest}
\alias{as.jointest}
\alias{plot.jointest}
\alias{p.adjust.jointest}
\title{Methods for jointest objects}
\usage{
\method{summary}{jointest}(object, ...)

\method{as}{jointest}(object, names_obj = NULL, ...)

\method{plot}{jointest}(object, p.values = c("raw", "adjusted"), mark_signif = 0.05)

\method{p.adjust}{jointest}(mods, method = "maxT", tail = 0, ...)
}
\arguments{
\item{object}{an object of class \code{jointest}.}

\item{...}{additional arguments to be passed}

\item{names_obj}{a vector of names, its length must be equal to the length of \code{object}}

\item{p.values}{use "raw" or "adjusted" (i.e. corrected for selective inference)}

\item{mark_signif}{a value between 0 and 1. The plot will mark the p-values smaller than \code{mark_signif} (0.05 by default). If equal to 0 or 1 nothing will be marked.}

\item{method}{any method implemented in \code{flip::flip.adjust} or a custom function. In the last case it must be a function that uses a matrix as input and returns a vector of adjusted p-values equal to the numbero fo columns of the inputed matrix.}
}
\description{
Methods for \code{jointest} objects. 
The following are methods to extract and manipulate relevant information from
a \code{jointest} object.
}
