% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jointest_package.R
\docType{package}
\name{jointest-package}
\alias{jointest-package}
\title{Robust Score Testing in GLMs, by Sign-Flip Contributions}
\description{
It provides robust tests for testing in GLMs, by sign-flipping score contributions. The tests are often robust against overdispersion, heteroscedasticity and, in some cases, ignored nuisance variables.
}
\author{
Livio Finos, Jelle Goeman and Jesse Hemerik, with contribution of Vittorio Giatti.
}
