% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/padjust.R
\name{p.adjust.fwer}
\alias{p.adjust.fwer}
\title{Correct p-values of a jointest object for multiple testing (FamilyWise Error Rate)}
\usage{
p.adjust.fwer(mods, method = "maxT", tail = 0, ...)
}
\arguments{
\item{mods}{object of class \code{jointest} (it can be a list of glm or flipscores converted in a \code{jointest} object using \code{as.jointest})}

\item{method}{any method implemented in \code{flip::flip.adjust} or a custom function. In the last case it must be a function that uses a matrix as input and returns a vector of adjusted p-values equal to the numbero fo columns of the inputed matrix.}

\item{tail}{= 0}

\item{...}{further parameters}
}
\description{
Correct p-values of a jointest object for multiple testing (FamilyWise Error Rate)
}
