% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine.R
\name{combine}
\alias{combine}
\title{(nonparametric) combination of jointest object}
\usage{
combine(mods, comb_funct = "maxT", combined = NULL, tail = 0)
}
\arguments{
\item{combined}{a) NULL = combines all coeffs with the same names along models, 
b) a vector of characters = it gives a combination for each stats with these names 
c) a list of arrays = each array has two columns reporting model (first column) 
and name of the coefficient to be combined.}
}
\description{
(nonparametric) combination of jointest object
}
