% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flip2sss.R
\name{flip2sss}
\alias{flip2sss}
\title{flip2sss}
\usage{
flip2sss(
  formula = NULL,
  data = NULL,
  cluster = NULL,
  summstats_within = NULL,
  ...
)
}
\arguments{
\item{formula}{A formula or a list of formulas. It can be a complete model as.formula or a list of formulas, one for each element produced by the function.}

\item{data}{The dataset to be used for fitting the model.}

\item{cluster}{A vector or a formula evaluated on the data that defines the clusters.}

\item{summstats_within}{A vector of summary statistics model within the data or a function with argument data.}

\item{...}{Other arguments passed to the `flipscores` function.}
}
\value{
A jointest object containing the model results. Note that the flipscores models for each coefficient within are also included in the jointest object.
}
\description{
This function fits a model based on the provided formula and data, accounting for clusters and summary statistics within the model.
}
\examples{
formula <- SALUTE ~ GENERE * TIME + PROV_BAMB + ETA_B_ARRIVO
data <- adolong
cluster <- adolong$SOGG
library(logistf)
temp=logistf::logistf(SALUTE ~ TIME, family = binomial(link = "logit"),data=data,control=logistf.control(maxit=100))
summstats_within <- 'logistf::logistf(SALUTE ~ TIME, family = binomial(link = "logit"),control=logistf::logistf.control(maxit=100))'
res <- flip2sss(formula, data, cluster, summstats_within)
attr(model.matrix(mods[[1]]),"assign")
}
