% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join_flipscores.R
\name{join_flipscores}
\alias{join_flipscores}
\title{jointest for flipscores objects}
\usage{
join_flipscores(
  mods,
  tested_coeffs = NULL,
  n_flips = 5000,
  score_type = "standardized",
  seed = NULL,
  statistics = "t"
)
}
\arguments{
\item{tested_coeff}{is a list of the same length of \code{mods}, each element of the list being a vector of
of names of tested coefficients. Alternatively, it can be a vector of names of tested coefficients, in this case, the tested coefficients are attributed to all models (when present). 
As a last option, it can be \code{NULL}, if so, all coefficients are tested.}
}
\description{
jointest for flipscores objects
}
\examples{
n=20
set.seed(123)
D=data.frame(X=rnorm(n),Z1=rnorm(n),Z2=rnorm(n))
D$Y=D$Z1+D$X+rnorm(n)
mod1=glm(Y~X+Z1+Z2,data=D)
mod2=glm(Y~X+poly(Z1,2)+Z2,data=D)
mod3=glm(Y~X+poly(Z1,2)+poly(Z2,2),data=D)
mod4=glm(Y~X+Z1+poly(Z2,2),data=D)
mods=list(mod1=mod1,mod2=mod2,mod3=mod3,mod4=mod4)
for(i in 1:length(mods))
mods[[i]]$call$data=eval(D)
library(jointest)
res=join_flipscores(mods,n_flips = 5000, score_type = "standardized" ,
      seed = 1, tested_coeffs = "X")
summary(res)
summary(combine(res))
res=p.adjust.fwer(res)
summary(res)
}
