% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join_flipscores.R
\name{join_flipscores}
\alias{join_flipscores}
\title{join tests from multiverse models}
\usage{
join_flipscores(mods, tested_coeffs = NULL, n_flips = 5000, 
score_type = "standardized", statistics = "t", seed=NULL, output_models =TRUE, ...)
}
\arguments{
\item{mods}{list of \code{glm}s or \code{flipscores}-object (or any other object that can be evaluated by \code{flipscores})}

\item{tested_coeffs}{list of the same length of \code{mods}, each element of the list being a vector of
names of tested coefficients. Alternatively, it can be a vector of names of tested coefficients, in this case, the tested coefficients are attributed to all models (when present). 
As a last option, it can be \code{NULL}, if so, all coefficients are tested.}

\item{n_flips}{number of flips, default 5000}

\item{score_type}{any valid type for \code{flipscores}, \code{"standardized"} is the default. see \code{\link[flipscores]{flipscores}} for more datails}

\item{statistics}{"t" is the only method implemented (yet). Any other value will not modify the score (a different statistic will only affect the multivariate inference, not the univariate one).}

\item{seed}{\code{NULL} by default. Use a number if you wanto to ensure replicability of the results}

\item{output_models}{\code{TRUE} by default. Should the \code{flipscores} model returned?}

\item{...}{any other further parameter.}
}
\value{
A \code{jointest} object, i.e., a list containing the following objects: 
\describe{
 \item{Tspace}{\code{data.frame} where rows represents the sign-flipping transformed (plus the identity one) test and columns the variables.}
 \item{summary_table}{\code{data.frame} containing for each model the estimated parameter(s), score(s), std error(s), test(s), partial correlation(s) and p-value(s).} 
 \item{mods}{List of \code{glm}s or \code{flipscores} objects.}
}
}
\description{
The function allows hypothesis testing across all plausible multiverse models
ensuring strong family-wise error rate control.
}
\examples{
library(jointest)
set.seed(123)

#Simulate data
n=20
D=data.frame(X=rnorm(n),Z1=rnorm(n),Z2=rnorm(n))
D$Y=D$Z1+D$X+rnorm(n)

# Run four glms abd combine it in a list
mod1=glm(Y~X+Z1+Z2,data=D)
mod2=glm(Y~X+poly(Z1,2)+Z2,data=D)
mod3=glm(Y~X+poly(Z1,2)+poly(Z2,2),data=D)
mod4=glm(Y~X+Z1+poly(Z2,2),data=D)
mods=list(mod1=mod1,mod2=mod2,mod3=mod3,mod4=mod4)

# Let us analyze the tests related to coefficient "X" and combine them
res=join_flipscores(mods,n_flips = 5000, seed = 1, tested_coeffs = "X")
summary(combine(res))
}
