% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{jointest-method}
\alias{jointest-method}
\alias{summary.jointest}
\alias{plot.jointest}
\title{Methods for jointest objects}
\usage{
\method{summary}{jointest}(object, ...)

\method{plot}{jointest}(mods, p.values = c("raw", "adjusted"), mark_signif = 0.05)
}
\arguments{
\item{object}{a jointest object}

\item{...}{additional arguments to be passed}
}
\description{
Methods for \code{jointest} objects. 
The following are methods to extract and manipulate relevant information from
a \code{jointest} object.
}
