% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jointest-methods.R
\docType{methods}
\name{jointest-methods}
\alias{jointest-methods}
\alias{print.jointest}
\alias{summary.jointest}
\alias{p.adjust.jointest}
\alias{plot.jointest}
\title{Methods for \code{jointest} objects}
\usage{
\method{print}{jointest}(x, ...)

\method{summary}{jointest}(object, ...)

\method{p.adjust}{jointest}(p, method = "maxT", ...)

\method{plot}{jointest}(x, ...)
}
\arguments{
\item{x}{an object of class \code{jointest}.}

\item{...}{additional arguments to be passed, i.e., \code{mark_signif} and
\code{p.values=c("raw","adjusted")}.}

\item{object}{an object of class \code{jointest}.}

\item{p}{an object of class \code{jointest}.}

\item{method}{any method implemented in \code{flip::flip.adjust} or 
a custom function. In the last case it must be a function that uses a matrix 
as input and returns a vector of adjusted p-values equal to the number of columns of the inputed matrix.}
}
\description{
Methods to extract and manipulate relevant information from
a \code{jointest} object.

\code{print} method for class "\code{jointest}".

\code{summary} method for class "\code{jointest}"

\code{p.adjust} method for class "\code{jointest}"

\code{plot} shows the p-values from multiverse models marked by
different shape for significance level defined in \code{mark_signif} argument
default is 0.05 and different colors for each coefficient.
}
\seealso{
\code{\link[flip]{flip.adjust}}
}
