% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_padjust.R
\name{p.adjust.fwer}
\alias{p.adjust.fwer}
\title{Correct p-values of a jointest object for multiple testing (FamilyWise Error Rate)}
\usage{
p.adjust.fwer(mods, method = "maxT", tail = 0)
}
\description{
Correct p-values of a jointest object for multiple testing (FamilyWise Error Rate)
}
